Not calcuating jiffies for scheduling delayed work
Turns out the logic for constructing the jiffies for calling
queue_delayed_work(9) was flawed. This commit resolves issue GH-10.

On Linux, the timer interrupt frequency (HZ) is configured and built
into the kernel as an immutable value. Supported values are 100, 250,
300, and 1000...though some others can exist.

According to time(8)[1], 1 / HZ = 1 jiffy.

So 1 second in jiffies should be HZ jiffies.

UNFORTUNATELY, given clock skew issues due to being stuck using
refined jiffies on my Alpine vm, it seems while we can translate
desired time delays (in seconds) to jiffies I tend to see it take
longer than expected for the delayed work to be executed.

This sort of begs for fixing the actual clock issues.

[1] https://linux.die.net/man/7/time