increase chances of finding the rtc device
Related to #14, on some systems we can't find the RTC device due to the
kernel config not having CONFIG_RTC_HCTOSYS_DEVICE defined, so try
falling back to the alternative CONFIG_RTC_SYSTOHC_DEVICE.

Not sure yet how to handle this in general, so this is a short term fix.
For instance, the driver will still schedule the work function for
synchronizing the clock and respond to sync commands. Wasteful.